chore: uppercase component files, index for root refs, uppercase component scss files
