+++ frontend/src/app/components/select/SelectOption.tsx
@@ -1,26 +1,26 @@
 /* Copyright (C) 2021 Vibey - All Rights Reserved */
 
 import React from 'react';
-import { components } from "react-select";
-import { Badge } from "@chakra-ui/react";
-import './select.scss';
+import { components } from 'react-select';
+import { Badge } from '@chakra-ui/react';
+import './Select.scss';
 
 /**
  * Custom option for tags to display the label and score
- * 
+ *
  * Need to figure out the correct type for the prop to get rid of any
- * 
+ *
  * @param props options props passed by react-select
  */
 export const SelectOption = (props: any) => {
-  return (
-    <components.Option {...props}>
-      <span className="text-grey-1">#</span> <span>{props.data.label}</span>{" "}
-      <Badge className="bg-primary rounded-sm ml-1 text-white">
-        {props.data.score}
-      </Badge>
-    </components.Option>
-  );
+ return (
+  <components.Option {...props}>
+   <span className="text-grey-1">#</span> <span>{props.data.label}</span>{' '}
+   <Badge className="bg-primary rounded-sm ml-1 text-white">
+    {props.data.score}
+   </Badge>
+  </components.Option>
+ );
 };
 
 /**
@@ -28,16 +28,15 @@ export const SelectOption = (props: any) => {
  * We don't have a 'tag not found' result as the not found
  * tags can be created. So the only time its going to be not found
  * is when the api is fetching the tags based on the input substring
- * 
+ *
  * Need to figure out the correct type for the prop to get rid of any
- * 
+ *
  * @param props options props passed by react-select
  */
 export const NoSelectOption = (props: any) => {
-  return (
-    <components.NoOptionsMessage {...props}>
-      <span>Start typing!</span>
-    </components.NoOptionsMessage>
-  );
+ return (
+  <components.NoOptionsMessage {...props}>
+   <span>Start typing!</span>
+  </components.NoOptionsMessage>
+ );
 };
-
