+++ frontend/src/app/hooks/useAxios.ts
@@ -1,6 +1,7 @@
 import axios from "axios";
 import { BASE_URL } from "app/static/url";
 
+// TODO add access token and refresh token to headers
 const instance = axios.create({
   baseURL: BASE_URL,
   withCredentials: true,
